Simplify interpolation of 2-D matrix transforms.
The decomposition of a transformation matrix into translations, rotation, scale and skew transforms is not unique. In some cases, the generalized 3-D decomposition does not align with the working draft for CSS transforms.

In the special case where the transforms being interpolated are both 2-D, a simplified model provides more restricted set of decomposition transforms with less computational overhead.

Bug: 797472
Change-Id: I2b8ba99fe02c2eef878d94f5dfaea55c39652759