Rework of "inherit the navigation initiator when navigating"
Spec: https://w3c.github.io/webappsec-csp/#initialize-document-csp

This is a rework and reland of https://chromium-review.googlesource.com/c/chromium/src/+/1314633

The initial patch got reverted because it did not pass the trusty leak
build checks. The issue was that holding to the initiator document as a
member in FrameLoader was holding said document alive unreasonably long.
Instead this rework holds a copy of the initiator's CSP.

Bug: 905301, 894228, 836148
Change-Id: Ic12c28d20c53def5d6753449c3c4da7de5242ca2