[EventTiming] Always buffer firstInput
Before, we'd buffer the firstInput entry only if it happened before
onload. This CL forces firstInput to always be buffered. This means
EventTiming code is called more in some tests, so the event hardware
timestamp is no longer guaranteed to be nonzero, and a couple of xr
tests require double RAF to avoid competing with the swap promises from
Event Timing.

Bug: 841224, 843184

Change-Id: I942b934c387798c100da4ecfff52affb66ab94e8